{{cookiecutter.project_slug}}/requirements/base.txt
# General
pytz==2018.7
python-slugify==1.2.6
Pillow==5.3.0
{%- if cookiecutter.use_compressor == "y" %}
rcssmin==1.0.6 {% if cookiecutter.windows == 'y' %} --install-option="--without-c-extensions"{% endif %}
{%- endif %}
argon2-cffi==18.3.0
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==4.1
{%- endif %}
redis>=2.10.5
{%- if cookiecutter.use_celery == "y" %}
celery==4.2.1
{%- if cookiecutter.use_docker == 'y' %}
flower==0.9.2 
{%- endif %}
{%- endif %}

# Django
django==2.0.9
django-environ==0.4.5
django-model-utils==3.1.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
{%- if cookiecutter.use_compressor == "y" %}
django-compressor==2.2
{%- endif %}
{%- if cookiecutter.use_sass_preprocessor == 'y' %}
libsass==0.15.1
django-sass-processor==0.7.2
{%- endif %}
django-redis==4.9.0
django-filter==2.0.0
django-tables2==2.0.2
tablib==0.12.1
{%- if cookiecutter.pdf_plugin %}
# xhtml2pdf==0.2.2
django-weasyprint==0.5.4
{%- endif %}

# Django REST Framework
djangorestframework==3.9.0
coreapi==2.3.3
